+++ src/main.ts
@@ -6,6 +6,7 @@ import { toggleLights } from './visual/lights.ts';
 import { octaveUp, octaveDown, transposeUpOne, transposeDownOne } from './audio/transposeOctave.ts';
 import { toggleModal } from './components/modal.ts';
 import './audio/recording.ts';
+import './core/time.ts';
 import { keyEventToBaseKey, registerKeyDown, registerKeyUp } from './core/keypress.ts';
 import { states } from './core/states.ts';
 // import './sheets/transcribe.ts';